[follow] New script
follow.js is a sipmle follow script which lets you quickly choose
links and elements using keyboard.  After pressing "f" key each
link and form form in a viewport is assigned a label which you can
type to choose the element.  This script is dvorak-optimised but
can be customised by editing the charset field.