[follow] Fixed label out of viewport bug + optimisation
The label was always attached to the top left corner of the element
meaning that if the top left corner of the element was outside of
the viewport the label was truncated or invisible.  This commit
fixes this issue.

Thi commit also introduces several optimisations like calculating
viewport only once per init, calculating bounding box once por
element, traversing arrays from the back, etc.