[follow] retMode introduced
retMode is a mode where linke are highlighted when they match
typed in label but are not automatically enterred.  You need to
press return key (hence name) to confirm.

What is great about this mode is that you can press space and
start typing another label and in this way select several links.
Pressing return when many links are selected will open them in
background (scripts have to be allowed to lower windows for this
to work).

In retMode labels may be of different lenghts -- mutch like
nubres, leading zeros are not dispalyed -- and so some links will
be easier to type in then in ret mode

The way of generating labels in normal mode has been changed as
well.  Labels are generated in reverse order so that labels with
the same first letter are not grouped together.  This may help
as user won't be distracted by all the elements around the one
she want to click.

This may lead to situations where a label's prefix is unique and
the last character(s) does not need to by typed in.  Ideally, in
the future such labels should be truncated.