[follow] A new labelling algorithm
A new labelling algorithm is now used for both the retMode and
normal mode.  This new scheme joins advantages of both old schemes.
The labels are not of equal length so some are one character shorter
then the others and no label is a prefix of another label.

Previously, in the normal mode labels were of equal lengths, however,
if prefix typed in uniquely identified label then it was chosen which
was kind of cool but also confusing.  With the new mode, such labels
are simply shorter which avoids confusion.

Moreover, the new scheme "prefers" links that appear first in the
document.links array and, if possible, those are the ones that
get shorter labels.