[follow] Link with the same href merger implemented
Previously, if two or more links had the same href they would
have separate labels which is a wast.  With this patch, such
links are assigned single labels.

Also, links that point to the page itself receive no label.

This has however limitations as the hrefs must be identical and
not merely point to the same URL (ie. “http://example.cam/” and
“/” are considered different links regardless if we are on the
“http://example.cam/whatever” website or somewhere else.