[backspace-blur] New script
backspace-blur is an UserJS which blurs form fields when
backspace is pressed.  Blurs af course means that focus
is moved out of them.  This means that all one-letter
shourtcuts start working again instead of genereting
characters in the field.  Obviously, the script kicks in
only if input fields are empty.