[tip] New script
tip.js is a script which shows information about hovered object in
lower left corner.  This is a replacement for tooltips and status
bar which has no delay (opposite to tooltips) and requires no
space on browser's user interface (so you can completely turn
status bar and address bar off).