[tip] Added list of attributes and show element info mode
Via a list of attributes user can easily specify which attributes
should be displayed.  An onClick attribute has also been added as
it may be pretty important (since it often changes the meaning of
href).  Other onClick handlers are not displayed though.

The element info mode makes the tooltip display information about
element that is hovered.  That is tag name, ID and class.